Start refactor
This module has many reported problems, such as fetching tokens too many
times, and it does not conform to the latest ecobee API format. The API
it's using is slated to be deprecated soon, so it needs to be updated.

Before I could start any of that, however, I needed to do some work on
culling unused portions of the module and generally getting the code to
be internally consistent. It looks to me like this was originally
adapted from a Nest module, so lots of the previous module's code was
still in place, albeit unused. There's a lot more to do here for
cleanup, modernization, and internal consistency, but I wanted to get
this checkpoint submitted so I'd have a safe base to build off of.

Essentially this deletes unused code, changes function() {} to use arrow
functions or member functions, reduces global variable usage, and
reformats the code to satisfy eslint and prettier.