Add svgs for each icon
This is a pull of various assets from ecobee's website for me to work
with while I find the best way to include them in a flexible way in the
module. Colors need to be set to white so they can be colorized by CSS,
and they probably all need to live directly in the html for the same
reason, not as separate assets.