Move the sensor refresh timer to happen on the server
Previously the client was responsible for keeping a timer of how often
to ask for new sensor data. This creates a problem when multiple clients
want to connect to the same server and can also cause unnecessary api
hits when restarting/refreshing the client side. Now the client sends
its desired update frequency to the server on startup and the server
schedules updates at the minimum desired rate of all connected clients
(or 3 minutes, whichever is longest, to keep the api side happy).

A happy side effect of this is that sensor update requests from new
clients get satisfied quickly since the server keeps the last response
data cached. Same goes for subsequent PIN requests when one is already
waiting.